+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ihatecode/react-splitter",
-  "version": "1.2.0",
+  "version": "2.0.0-alpha.0",
   "description": "A split pane component written in React.",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
