+++ src/components/style.css
@@ -2,43 +2,56 @@
     width: 100%;
     height: 100%;
     display: flex;
+
+    --ihc-splitter-size: 1px;
+    --ihc-splitter-color: #eee;
 }
 
 .ihc-splitter-vertical {
     flex-direction: column;
 }
 
-.ihc-splitter-bordered {
-    border: 1px solid #eee;
-}
-
 .ihc-splitter-resizing {
     user-select: none;
 }
 
+.ihc-splitter-horizontal.ihc-splitter-resizing {
+    cursor: col-resize;
+}
+
+.ihc-splitter-vertical.ihc-splitter-resizing {
+    cursor: row-resize;
+}
+
 .ihc-splitbar {
     position: relative;
-    background-color: #eee;
+    background-color: var(--ihc-splitter-color);
 }
 
 .ihc-splitbar-vertical {
-    height: 1px;
+    height: var(--ihc-splitter-size);
+    min-height: var(--ihc-splitter-size);
     width: 100%;
     cursor: row-resize;
 }
 
 .ihc-splitbar-horizontal {
-    width: 1px;
+    width: var(--ihc-splitter-size);
+    min-width: var(--ihc-splitter-size);
     height: 100%;
     cursor: col-resize;
 }
 
+.ihc-splitbar-nonresizable {
+    cursor: default;
+}
+
 .ihc-splitter-item {
-    width: 100%;
-    height: 100%;
     overflow: hidden;
+    transition: flex-basis .15s ease-in-out;
 }
 
 .ihc-splitter-resizing>.ihc-splitter-item {
     will-change: flex-basis;
+    transition: flex-basis 0s ease-in-out;
 }
